Disable X11 with --disable-x11.
