+++ interface/VideoEncoderInterface.h
@@ -21,9 +21,14 @@
 
 #ifndef VIDEO_ENCODER_INTERFACE_H_
 #define VIDEO_ENCODER_INTERFACE_H_
+#ifdef HAVE_CONFIG_H
+#include "config.h"
+#endif
 
 #include "VideoEncoderDefs.h"
+#ifdef __ENABLE_X11__
 #include <X11/Xlib.h>
+#endif
 #undef None // work around for compile in chromeos
 
 namespace YamiMediaCodec{
